Remove references from base jax-rs APP to avoid CNFE
The base jaxrs app is removed from deployment in rest-client module
therefore it cannot be referenced from other resource classes.
